requirements.txt
pydantic
pytest
python-dateutil
aiofiles
aiocsv
pytest-asyncio
